refactor(runner): extract job provider trait to decouple lifecycle from transport
Move Ably push + HTTP poll + REST API claim/complete logic from start.rs
into a new provider module with a clean trait interface:

- JobProvider trait: next_job() for discovery+claim, complete() for
  reporting results — hides all transport details from the main loop
- ApiProvider: production implementation wrapping Ably + poll + REST API
- RetryState<H>: generic backoff/retry shared between provider (Ably) and
  main loop (mitmproxy)

The main loop in start.rs shrinks from ~993 to ~560 lines, with a simplified
tokio::select! that just calls provider.next_job() instead of managing 7
interleaved branches for Ably/poll/claim/complete.

Closes #3155
