+++ crates/runner/src/provider/mod.rs
@@ -1,0 +1,191 @@
+//! Job provider trait and shared retry utilities.
+//!
+//! The [`JobProvider`] trait abstracts job lifecycle (discovery, claiming,
+//! completion reporting) so different transports can be plugged in without
+//! changing the executor or main loop.
+
+mod api;
+
+pub use api::ApiProvider;
+
+use std::time::{Duration, Instant};
+
+use uuid::Uuid;
+
+use crate::types::ExecutionContext;
+
+// ---------------------------------------------------------------------------
+// Trait
+// ---------------------------------------------------------------------------
+
+/// Abstraction over job lifecycle — discovery, claiming, and completion reporting.
+///
+/// The runner main loop calls [`next_job()`](JobProvider::next_job) to get work
+/// and [`complete()`](JobProvider::complete) to report results. All transport
+/// details (Ably push, HTTP poll, WebSocket, etc.) are hidden behind this trait.
+#[async_trait::async_trait]
+pub trait JobProvider: Send + Sync {
+    /// Wait for the next available job. Returns `None` on shutdown signal.
+    ///
+    /// Implementations handle discovery (push/poll), claiming, and retry logic
+    /// internally. The returned [`ExecutionContext`] is ready for execution.
+    async fn next_job(&self) -> Option<ExecutionContext>;
+
+    /// Report job completion. Called concurrently from spawned executor tasks.
+    ///
+    /// Implementations manage auth tokens and retry logic internally.
+    async fn complete(&self, run_id: Uuid, exit_code: i32, error: Option<&str>);
+}
+
+// ---------------------------------------------------------------------------
+// RetryState (shared with start.rs for mitmproxy)
+// ---------------------------------------------------------------------------
+
+/// Groups the backoff / retry state for a restartable background task.
+pub(crate) struct RetryState<H> {
+    pub(crate) handle: Option<H>,
+    pub(crate) restart_at: Option<Instant>,
+    pub(crate) backoff: Duration,
+    backoff_initial: Duration,
+    backoff_max: Duration,
+    pub(crate) consecutive_failures: u32,
+    /// `None` = retry forever (Ably), `Some(n)` = circuit breaker (mitm).
+    max_failures: Option<u32>,
+}
+
+impl<H> RetryState<H> {
+    pub(crate) fn new(initial: Duration, max: Duration, max_failures: Option<u32>) -> Self {
+        Self {
+            handle: None,
+            restart_at: None,
+            backoff: initial,
+            backoff_initial: initial,
+            backoff_max: max,
+            consecutive_failures: 0,
+            max_failures,
+        }
+    }
+
+    /// Schedule a restart after the current backoff delay.
+    pub(crate) fn schedule(&mut self) {
+        self.restart_at = Some(Instant::now() + self.backoff);
+    }
+
+    /// Reset backoff and failure count after a successful restart.
+    pub(crate) fn on_success(&mut self) {
+        self.backoff = self.backoff_initial;
+        self.consecutive_failures = 0;
+    }
+
+    /// Record a failure, double the backoff (capped), and schedule a retry.
+    /// Returns `false` if the circuit breaker has tripped.
+    #[must_use]
+    pub(crate) fn on_failure(&mut self) -> bool {
+        self.consecutive_failures += 1;
+        if let Some(max) = self.max_failures
+            && self.consecutive_failures >= max
+        {
+            return false;
+        }
+        self.schedule();
+        self.backoff = (self.backoff * 2).min(self.backoff_max);
+        true
+    }
+
+    /// `true` if the restart timer has fired and no task is in flight.
+    pub(crate) fn timer_ready(&self) -> bool {
+        self.handle.is_none() && self.restart_at.is_some_and(|at| Instant::now() >= at)
+    }
+
+    /// Clear the timer after spawning a restart task.
+    pub(crate) fn clear_timer(&mut self) {
+        self.restart_at = None;
+    }
+}
+
+/// Sleep until a restart timer fires, or pend forever if none is scheduled.
+///
+/// Free function (not a method) so the borrow on `restart_at` is disjoint
+/// from `&mut retry.handle` inside `tokio::select!`.
+pub(crate) async fn sleep_until_retry(restart_at: &Option<Instant>) {
+    match restart_at {
+        Some(at) => tokio::time::sleep_until(tokio::time::Instant::from_std(*at)).await,
+        None => std::future::pending().await,
+    }
+}
+
+/// Await a background retry task, or pend forever if none is running.
+pub(crate) async fn recv_retry<T, E: std::fmt::Display>(
+    handle: &mut Option<tokio::task::JoinHandle<Result<T, E>>>,
+) -> Result<T, String> {
+    match handle {
+        Some(h) => {
+            let result = match h.await {
+                Ok(Ok(val)) => Ok(val),
+                Ok(Err(e)) => Err(e.to_string()),
+                Err(e) => Err(format!("retry task panicked: {e}")),
+            };
+            *handle = None;
+            result
+        }
+        None => std::future::pending().await,
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Tests
+// ---------------------------------------------------------------------------
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn retry_state_new_defaults() {
+        let rs: RetryState<()> =
+            RetryState::new(Duration::from_secs(1), Duration::from_secs(60), None);
+        assert_eq!(rs.consecutive_failures, 0);
+        assert!(rs.restart_at.is_none());
+        assert!(rs.handle.is_none());
+    }
+
+    #[test]
+    fn retry_state_on_failure_doubles_backoff() {
+        let mut rs: RetryState<()> =
+            RetryState::new(Duration::from_secs(1), Duration::from_secs(60), None);
+        let cont = rs.on_failure();
+        assert!(cont);
+        assert_eq!(rs.consecutive_failures, 1);
+        assert_eq!(rs.backoff, Duration::from_secs(2));
+        assert!(rs.restart_at.is_some());
+    }
+
+    #[test]
+    fn retry_state_on_failure_caps_at_max() {
+        let mut rs: RetryState<()> =
+            RetryState::new(Duration::from_secs(1), Duration::from_secs(4), None);
+        rs.backoff = Duration::from_secs(4);
+        let _ = rs.on_failure();
+        assert_eq!(rs.backoff, Duration::from_secs(4));
+    }
+
+    #[test]
+    fn retry_state_circuit_breaker() {
+        let mut rs: RetryState<()> =
+            RetryState::new(Duration::from_secs(1), Duration::from_secs(60), Some(3));
+        rs.consecutive_failures = 2;
+        let cont = rs.on_failure();
+        assert!(!cont, "circuit breaker should trip at max_failures");
+    }
+
+    #[test]
+    fn retry_state_on_success_resets() {
+        let mut rs: RetryState<()> =
+            RetryState::new(Duration::from_secs(1), Duration::from_secs(60), None);
+        rs.backoff = Duration::from_secs(32);
+        rs.consecutive_failures = 10;
+        rs.on_success();
+        assert_eq!(rs.backoff, Duration::from_secs(1));
+        assert_eq!(rs.consecutive_failures, 0);
+    }
+}
